Add trigger for block placements and block placement of SpecialItems
Blocking of placement is necessary as some blocks (stairs) seem to not
trigger the interact event. In that case we want the special item to
just not do anything. If placement of blocks is what the item should be
able to do then the no-cancel action can be used for the place trigger.